+++ src/object.h
@@ -741,6 +741,7 @@ public:
     NONE,
     ITER,
     SORT_WITH,
+    HISTORY,
   };
 
 private: